ref: split modelCleaning logic
